updated code generation and refactored test utils
