+++ internals/generators/ams-container/actions.test.js.hbs
@@ -1,3 +1,4 @@
+import { testActionCreator } from 'test/utils';
 
 import {
   defaultAction,
@@ -7,12 +8,8 @@ import {
 } from './constants';
 
 describe('{{ properCase name }} actions', () => {
-  describe('Default Action', () => {
-    it('has a type of DEFAULT_ACTION', () => {
-      const expected = {
-        type: DEFAULT_ACTION,
-      };
-      expect(defaultAction()).toEqual(expected);
-    });
+  it('should dispatch defaultAction action', () => {
+    const payload = {};
+    testActionCreator(defaultAction, DEFAULT_ACTION, payload);
   });
 });
