+++ internals/generators/ams-container/reducer.test.js.hbs
@@ -1,8 +1,22 @@
 import { fromJS } from 'immutable';
 import {{ camelCase name }}Reducer from './reducer';
 
+import {
+  DEFAULT_ACTION,
+} from './constants';
+
 describe('{{ camelCase name }}Reducer', () => {
   it('returns the initial state', () => {
     expect({{ camelCase name }}Reducer(undefined, {})).toEqual(fromJS({}));
+  });
+
+  it('should DEFAULT_ACTION', () => {
+    expect(
+      {{ camelCase name }}Reducer(fromJS({}), {
+        type: DEFAULT_ACTION
+      }).toJS()
+    ).toEqual({
+      foo: 'bar'
+    });
   });
 });
