+++ internals/generators/ams-model/actions.test.js.hbs
@@ -1,3 +1,4 @@
+import { testActionCreator } from 'test/utils';
 
 import {
   defaultAction,
@@ -6,13 +7,9 @@ import {
   DEFAULT_ACTION,
 } from './constants';
 
-describe('{{ camelCase name }} actions', () => {
-  describe('Default Action', () => {
-    it('has a type of DEFAULT_ACTION', () => {
-      const expected = {
-        type: DEFAULT_ACTION,
-      };
-      expect(defaultAction()).toEqual(expected);
-    });
+describe('{{ properCase name }} actions', () => {
+  it('should dispatch defaultAction action', () => {
+    const payload = {};
+    testActionCreator(defaultAction, DEFAULT_ACTION, payload);
   });
 });
