+++ package.json
@@ -205,6 +205,7 @@
     "collectCoverageFrom": [
       "src/**/*.{js,jsx}",
       "!src/**/*.test.{js,jsx}",
+      "!src/test/**/*.{js,jsx}",
       "!src/*/RbGenerated*/*.{js,jsx}",
       "!src/app.js",
       "!src/global-styles.js",
