+++ src/containers/App/actions.test.js
@@ -1,4 +1,4 @@
-import { testActionCreator } from '../../../internals/testing/test-utils';
+import { testActionCreator } from 'test/utils';
 
 import {
   AUTHENTICATE_USER,
